moved Match detection alerts and tests to Alert+Platform and AlertPlatformTests
